+++ lib/simple_crud/cache_helpers.rb
@@ -7,5 +7,11 @@ module SimpleCrud
     def fetch_cached(key, ttl, &block)
       Rails.cache.fetch(key, expires_in: ttl, &block)
     end
+
+    def expire_simple_crud_cache(action, path: request.fullpath)
+      Rails.cache.delete(
+        SimpleCrud::ActionContext.cache_key_for(self.class.simple_crud_controller_model, action, path)
+      )
+    end
   end
 end