+++ lib/simple_crud/new_context.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
 module SimpleCrud
-  class NewContext < FormContext
+  class NewContext < ActionContext
     private
 
     def run
       record = build_record
       maybe_authorize(record)
-      render_form_action(record, :new)
+      render_record(record, :new)
     end
   end
 end